Refactor component versions to use join
